requirement.txt
Flask
opencv-python
easyocr
Pillow
numpy